owly/Owly/Scripts/requirements.txt
openai
cloudinary
requests
pytesseract
Pillow
spacy
torch
transformers
